+++ GameProject/Engine/InputHandler.h
@@ -8,7 +8,7 @@ private:
 	GLFWwindow* window;
 
 	// Used for mouse travel distance since last callback calculation
-	static int lastPosX, lastPosY;
+	static double lastPosX, lastPosY;
 
 	static void keyCallback(GLFWwindow* window, int key, int scancode, int action, int mods);
 	static void mouseMoveCallback(GLFWwindow* window, double xpos, double ypos);
